+++ src/apps/ScopeChangeRequest/Components/STID/Components/AdvancedDocumentSearch.tsx
@@ -1,8 +1,8 @@
-import React, { Fragment, useEffect, useReducer, useState } from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 import { Button, Icon, Progress, Scrim, SingleSelect, TextField } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
 import { TypedSelectOption } from '../../../Api/Search/searchType';
-import { StidTypes } from '../../../Api/Search/STID/searchStid';
+import { StidTypes } from '../../../Types/STID/STIDTypes';
 import { useHttpClient } from '../../../../../Core/Client/Hooks/useApiClient';
 import { getDocumentsByTag } from '../../../Api/STID/getDocumentsByTag';
 import { Result } from './Results';
@@ -32,6 +32,7 @@ export const AdvancedDocumentSearch = ({
 
     const { search } = useReferencesSearch();
 
+    const [isLoading, setIsLoading] = useState<boolean>(false);
     //controls
     const [isOpen, setIsOpen] = useState(false);
     const [searchText, setSearchText] = useState<string | undefined>();
@@ -44,11 +45,6 @@ export const AdvancedDocumentSearch = ({
     function checkDuplicate(x: TypedSelectOption): boolean {
         return documents.map((x) => x.value).includes(x.value);
     }
-
-    const [
-        { addDuplicate, documentsLoading, tagsLoading, hasErrored, tagContainsNoDocuments },
-        dispatch,
-    ] = useReducer(tasksReducer, initial);
 
     const popResult = (value: string) => {
         setResults((prev) => prev.filter((x) => x.value !== value));
@@ -104,7 +100,6 @@ export const AdvancedDocumentSearch = ({
     async function handleStidTagSelected(item: TypedSelectOption) {
         const documents = await getDocumentsByTag('JCA', item.value, STID);
         if (documents.length === 0) {
-            dispatch({ type: 'setTagContainsNoDocuments', value: true });
             popResult(item.value);
         } else {
             setSubResults({
@@ -126,61 +121,17 @@ export const AdvancedDocumentSearch = ({
         setResults([]);
         setSubResults(undefined);
         setSearchText(undefined);
-        dispatch({ type: 'reset', value: true });
     }
 
     const fetchResults = async (inputValue: string) => {
+        setIsLoading(true);
         //Aborts previous api call
         abort();
         if (!referenceType) return;
-
         const data = await search(inputValue, referenceType, getSignal());
 
         setResults(data);
-
-        // switch (referenceType) {
-        //     case 'commpkg': {
-        //         const commPkgs = await searchPcs(inputValue, 'commpkg', getSignal());
-        //         setResults(commPkgs);
-        //         break;
-        //     }
-
-        //     case 'system': {
-        //         const commPkgs = await searchPcs(inputValue, 'system', getSignal());
-        //         setResults(commPkgs);
-        //         break;
-        //     }
-
-        //     case 'tag': {
-        //         const commPkgs = await searchPcs(inputValue, 'tag', getSignal());
-        //         setResults(commPkgs);
-        //         break;
-        //     }
-
-        //     case 'discipline': {
-        //         const commPkgs = await searchPcs(inputValue, 'discipline', getSignal());
-        //         setResults(commPkgs);
-        //         break;
-        //     }
-
-        //     case 'area': {
-        //         const commPkgs = await searchPcs(inputValue, 'area', getSignal());
-        //         setResults(commPkgs);
-        //         break;
-        //     }
-
-        //     case 'stidtag': {
-        //         const tagsResult = await searchStid(inputValue, 'stidtag', getSignal());
-        //         setResults(tagsResult);
-        //         break;
-        //     }
-
-        //     case 'document': {
-        //         const documentsResult = await searchStid(inputValue, 'document', getSignal());
-        //         setResults(documentsResult);
-        //         break;
-        //     }
-        // }
+        setIsLoading(false);
     };
 
     return (
@@ -231,6 +182,8 @@ export const AdvancedDocumentSearch = ({
                                 label="Reference type"
                                 items={referenceTypes}
                                 handleSelectedItemChange={(change) => {
+                                    abort();
+                                    setSearchText('');
                                     if (!change.selectedItem) {
                                         setReferenceType(undefined);
                                     } else {
@@ -247,7 +200,7 @@ export const AdvancedDocumentSearch = ({
                                 value={searchText}
                                 inputIcon={
                                     <>
-                                        {documentsLoading || tagsLoading ? (
+                                        {isLoading ? (
                                             <Progress.Dots color="primary" />
                                         ) : (
                                             <Icon name="search" />
@@ -263,23 +216,6 @@ export const AdvancedDocumentSearch = ({
                                     }
                                 }}
                             />
-                        </div>
-                        <div style={{ height: '12px' }}>
-                            {tagContainsNoDocuments && (
-                                <div style={{ color: 'red' }}>This tag contains no documents</div>
-                            )}
-                            {addDuplicate && (
-                                <div style={{ color: 'red' }}>
-                                    This document has already been added
-                                </div>
-                            )}
-                            {(tagsLoading || documentsLoading) && (
-                                <div>
-                                    {tagsLoading && <div>Searching tags</div>}
-                                    {documentsLoading && <div>Searching documents</div>}
-                                </div>
-                            )}
-                            {hasErrored && <div>Stid call failed</div>}
                         </div>
 
                         {subResults ? (
@@ -309,64 +245,3 @@ export const AdvancedDocumentSearch = ({
         </Fragment>
     );
 };
-
-const initial: NotifyEvents = {
-    addDuplicate: false,
-    documentsLoading: false,
-    hasErrored: false,
-    tagContainsNoDocuments: false,
-    tagsLoading: false,
-};
-
-interface NotifyEvents {
-    tagContainsNoDocuments: boolean;
-    tagsLoading: boolean;
-    documentsLoading: boolean;
-    hasErrored: boolean;
-    addDuplicate: boolean;
-}
-
-type MutableStates =
-    | 'setAddDuplicate'
-    | 'setDocumentsLoading'
-    | 'setHasErrored'
-    | 'setTagContainsNoDocuments'
-    | 'setTagsLoading'
-    | 'reset';
-
-interface Actions {
-    type: MutableStates;
-    value: boolean;
-}
-
-function tasksReducer(state: NotifyEvents, action: Actions): NotifyEvents {
-    switch (action.type) {
-        case 'setAddDuplicate': {
-            return { ...state, addDuplicate: action.value };
-        }
-
-        case 'setDocumentsLoading': {
-            return { ...state, documentsLoading: action.value };
-        }
-
-        case 'setHasErrored': {
-            return { ...state, hasErrored: action.value };
-        }
-
-        case 'setTagContainsNoDocuments': {
-            return { ...state, tagContainsNoDocuments: action.value };
-        }
-
-        case 'setTagsLoading': {
-            return { ...state, tagsLoading: action.value };
-        }
-
-        case 'reset': {
-            return initial;
-        }
-
-        default: {
-            throw Error('Unknown action: ' + action.type);
-        }
-    }
-}
