+++ src/apps/ScopeChangeRequest/Components/STID/index.ts
@@ -1,2 +1,2 @@
-export * from './Components/StidSelector';
+export * from './Components/AdvancedDocumentSearch';
 export * from './Components/StidDocument';
