+++ src/apps/ScopeChangeRequest/Types/STID/STIDTypes.ts
@@ -1,0 +1,1 @@
+export type StidTypes = 'document' | 'stidtag';
